Initial logic to remove SQL-style comments.
In SQL, comments can be noted in two different ways. This logic
removes the type I use the most which is two dashes "--". Much like a
"#" in R, the double dashes can appear anywhere in a line. From that
point on, the line is considered to be a comment.

Also adds a simple gsub to remove the resulting double spaces in the
returned query which makes it easier for users to read the query.