sqldoc() now handles the SQL version of Roxygen comments.
SQL Roxygen comments are now "--'" which is directly inspired by
Roxygen comments which are "#'".

This needs more testing, but it appears to handle the following query
correctly:

--' This query pulls some_col from table foo.
--' This query is effectively a bullshit test of sqldoc().
--' @param bar This is my example parameter.
select some_col
-- I should not be returned by sqldoc().
from foo
where :bar: = 6 -- end-of-line comment. I should not be returned by sqldoc().
;

When this query is parsed by this version of sqldoc, the first three
comment lines ARE returned. The last two are not, which is exactly
what we want.